module pivot
structure is deprecated.
long live mapFolder!
