+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = require('./lib/structure');
+module.exports = require('./lib/map-folder');
