+++ lib/constants.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+    FOLDER: 0,
+    FILE: 1,
+    // SYMLINK: 2
+};
