+++ lib/create-done.js
@@ -1,0 +1,17 @@
+'use strict';
+
+module.exports = function (folderMapObj, maxCount, endCallBack) {
+    let doneCount = 0;
+    
+    return function (err) {
+        if (err) {
+            endCallBack(err, null);
+        }
+
+        doneCount++;
+
+        if (doneCount === maxCount) {
+            endCallBack(null, folderMapObj);
+        }
+    };
+};
