+++ lib/handle-entry-stat.js
@@ -1,0 +1,44 @@
+'use strict';
+
+const getEntryType = require('./get-entry-type');
+const getOctalMode = require('./get-octal-mode');
+
+
+let mapFolder;
+
+const FILE = require('./constants').FILE;
+
+module.exports = function (folderMapObj, entryTempObj, done) {
+    // TODO: fix circular dependency issue 
+    /* 
+        mapFolder(!) requires forEachEntry 
+        forEachEntry requires handleEntryStat 
+        handleEntryStat requires mapFolder(!)
+    */
+    mapFolder = mapFolder || require('./map-folder');
+
+    const {
+        name,
+        path,
+        stat
+    } = entryTempObj;
+    
+    const type = getEntryType(stat);
+
+    if (type === FILE) {
+        folderMapObj.entries[name] = {
+            path,
+            type: FILE,
+            mode: getOctalMode(stat.mode)
+        };
+
+        done(null, folderMapObj);
+    }
+    else { // FOLDER
+        mapFolder(path, (mapErr, result) => {
+            folderMapObj.entries[name] = result;
+
+            done(null, folderMapObj);
+        });
+    }
+};
