--- lib/structure.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = {
-    toJson: require('./to-json'),
-};
