+++ test/map-folder/a/index.js
