+++ test/map-folder/b/index.js
