+++ test/map-folder/index.js
